mcp_servers/python/servers/MCP-FARGATE/mcp-server-with-fargate/src/requirements.txt
fastapi
uvicorn
boto3
fastapi_mcp
python-dotenv
